Implementation of "CompetenceDAO" and associated test

+++ sadoc-parent/sadoc-dao-parent/sadoc-dao-commun/src/main/java/fr/univartois/ili/sadoc/dao/services/ICompetenceDAO.java
@@ -18,7 +18,7 @@ public interface ICompetenceDAO {
 	 * @return Competence
 	 * return the corresponding competence
 	 */
-	Competence findCompetenceById(int id);
+	Competence findCompetenceById(long id);
 
 	/**
 	 * Insert the competence in DB
